Renamespace reindex_request_creator to match the processor
