+++ reindexer/reindex_request_creator/src/main/scala/uk/ac/wellcome/platform/reindex/creator/models/ReindexJob.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.reindex_request_creator.models
+package uk.ac.wellcome.platform.reindex.creator.models
 
 /** A request to identify all the shards in the table that need reindexing.
   *
