+++ reindexer/reindex_request_creator/src/main/scala/uk/ac/wellcome/platform/reindex/creator/services/NotificationSender.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.reindex_request_creator.services
+package uk.ac.wellcome.platform.reindex.creator.services
 
 import com.google.inject.Inject
 import uk.ac.wellcome.messaging.sns.SNSWriter
